Refactor login package to align with security module
Moved the login package into the security module for better organization and cohesion. Updated imports and references across the codebase to reflect the new structure. Added a `codegen` step to the Makefile to support the changes.
