+++ server/base_server.go
@@ -4,8 +4,8 @@ import (
 	"github.com/gorilla/mux"
 	apicontext "github.com/softwareplace/http-utils/context"
 	errorhandler "github.com/softwareplace/http-utils/error"
-	"github.com/softwareplace/http-utils/login"
 	"github.com/softwareplace/http-utils/security"
+	"github.com/softwareplace/http-utils/security/login"
 	"github.com/softwareplace/http-utils/security/principal"
 	"github.com/softwareplace/http-utils/security/secret"
 	"net/http"
